+++ lib/src/extensions/flutter_extensions/navigation.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 extension FHUNavigationStateExtensions on State {
@@ -40,6 +39,11 @@ extension FHUNavigationStatelessExtensions on StatelessWidget {
 }
 
 extension FHUNavigatorExtension on BuildContext {
+  NavigatorState navigator({bool rootNavigator = false}) =>
+      Navigator.of(this, rootNavigator: rootNavigator);
+
+  ModalRoute<T>? modalRoute<T extends Object?>() => ModalRoute.of<T>(this);
+
   void popPage<T extends Object?>([T? result]) =>
       Navigator.pop<T>(this, result);
 
@@ -49,13 +53,10 @@ extension FHUNavigatorExtension on BuildContext {
   void popRoot<T extends Object?>([T? result]) =>
       navigator(rootNavigator: true).pop<T>(result);
 
-  NavigatorState navigator({bool rootNavigator = false}) =>
-      Navigator.of(this, rootNavigator: rootNavigator);
-
-  ///  just call this [canPop()] method and it would return true if this route can be popped and false if it’s not possible.
+  /// Just call this [canPop()] method and it would return true if this route can be popped and false if it’s not possible.
   bool get canPop => Navigator.canPop(this);
 
-  /// performs a simple [Navigator.push] action with given [route]
+  /// Performs a simple [Navigator.push] action with given [route]
   Future<T?> pushPage<T extends Object?>(
     Widget screen, {
     RouteSettings? settings,
@@ -63,7 +64,7 @@ extension FHUNavigatorExtension on BuildContext {
     bool fullscreenDialog = false,
     bool allowSnapshotting = true,
   }) async =>
-      Navigator.of(this).push<T>(
+      navigator().push<T>(
         MaterialPageRoute(
           builder: (_) => screen,
           settings: settings,
@@ -73,7 +74,7 @@ extension FHUNavigatorExtension on BuildContext {
         ),
       );
 
-  /// performs a simple [Navigator.pushReplacement] action with given [route]
+  /// Performs a simple [Navigator.pushReplacement] action with given [route]
   Future<T?> pReplacement<T extends Object?, TO extends Object?>(
     Widget screen, {
     bool allowSnapshotting = true,
@@ -81,7 +82,7 @@ extension FHUNavigatorExtension on BuildContext {
     bool maintainState = true,
     bool fullscreenDialog = false,
   }) async =>
-      Navigator.of(this).pushReplacement<T, TO>(
+      navigator().pushReplacement<T, TO>(
         MaterialPageRoute(
           allowSnapshotting: allowSnapshotting,
           builder: (_) => screen,
@@ -91,7 +92,7 @@ extension FHUNavigatorExtension on BuildContext {
         ),
       );
 
-  /// perform push and remove route
+  /// Perform push and remove route
   Future<T?> pAndRemoveUntil<T extends Object?>(
     Widget screen, {
     bool allowSnapshotting = true,
@@ -100,7 +101,7 @@ extension FHUNavigatorExtension on BuildContext {
     bool fullscreenDialog = false,
     bool routes = false,
   }) async =>
-      Navigator.of(this).pushAndRemoveUntil<T>(
+      navigator().pushAndRemoveUntil<T>(
         MaterialPageRoute(
           allowSnapshotting: allowSnapshotting,
           builder: (_) => screen,
@@ -111,69 +112,86 @@ extension FHUNavigatorExtension on BuildContext {
         (Route<dynamic> route) => routes,
       );
 
-  /// perform push and remove route with routeName
+  /// Perform push and remove route with routeName
   Future<T?> pNamedAndRemoveUntil<T extends Object?>(
     String screenName,
     RoutePredicate predicate, {
     bool allowSnapshotting = true,
     Object? arguments,
   }) async =>
-      Navigator.of(this).pushNamedAndRemoveUntil<T>(
+      navigator().pushNamedAndRemoveUntil<T>(
         screenName,
-        (route) => false,
+        predicate,
         arguments: arguments,
       );
 
-  /// perform push with routeName
+  /// Perform push with routeName
   Future<T?> pNamed<T extends Object?>(
     String screenName, {
     Object? arguments,
   }) async =>
-      Navigator.of(this).pushNamed<T>(screenName, arguments: arguments);
+      navigator().pushNamed<T>(screenName, arguments: arguments);
 
-  /// perform replace with routeName
+  /// Perform replace with routeName
   Future<T?> pReplacementNamed<T extends Object?, TO extends Object?>(
     String screenName, {
     Object? arguments,
     TO? result,
   }) =>
-      Navigator.of(this).pushReplacementNamed<T, TO>(
+      navigator().pushReplacementNamed<T, TO>(
         screenName,
         arguments: arguments,
         result: result,
       );
 
-  /// perform replace with routeName
+  /// Perform pop until a specific route name
   void popUntil(String screenName) =>
-      Navigator.of(this).popUntil(ModalRoute.withName(screenName));
+      navigator().popUntil(ModalRoute.withName(screenName));
 
   /// This method calls [Navigator.maybePop] to attempt to pop the current route if possible.
   /// It is useful when you want to handle back navigation and check if the route can be popped
   /// without causing any errors.
   Future<void> maybePop<T extends Object?>([T? result]) async =>
-      Navigator.of(this).maybePop<T>(result);
+      navigator().maybePop<T>(result);
 
-  /// This method programmatically and recursively dismisses any active pop-up elements like dialogs, modal bottom sheets,
-  ///   and Cupertino modal popups.
-  ///   It checks for the types [PopupRoute], [DialogRoute], [RawDialogRoute], [ModalBottomSheetRoute],
-  ///   and [CupertinoModalPopupRoute] to determine if a pop-up is currently displayed and closes it.
-  ///   If multiple pop-ups are stacked, the method will recursively close all of them.
-  void dismissActivePopup() {
+  /// Recursively dismisses active pop-up elements within the app.
+
+  /// This method automatically closes pop-up elements such as dialogs, modal bottom sheets,
+  /// and Cupertino modal popups. It does this by programmatically navigating back in the
+  /// widget tree until no pop-up routes remain.
+
+  /// The following types of routes are considered pop-ups:
+  ///  * `PopupRoute`
+  ///  * `DialogRoute`
+  ///  * `RawDialogRoute`
+  ///  * `ModalBottomSheetRoute`
+  ///  * `CupertinoModalPopupRoute`
+
+  /// **Parameters:**
+  ///  * `usePopUntil` (Optional, default: `true`):
+  ///    - If `true`, the `Navigator.popUntil` method is used, which efficiently dismisses
+  ///      all pop-ups in a single operation.
+  ///    - If `false`, the `Navigator.pop` method is used recursively, closing one pop-up at a time.
+  ///      This might be necessary in scenarios where `popUntil` is not supported or causes unexpected behavior.
+
+  /// **Usage:**
+
+  /// ```dart
+  ///  // Dismiss all active pop-ups using the most efficient method available.
+  ///  dismissActivePopup();
+  ///
+  ///  // Dismiss pop-ups one by one (might be needed for compatibility reasons).
+  ///  dismissActivePopup(usePopUntil: false);
+  /// ```
+  void dismissActivePopup({bool usePopUntil = true}) {
     try {
-      // Get the current route
-      final currentRoute = ModalRoute.of(this);
       final navigator = this.navigator(rootNavigator: true);
-      // Check if the current route is a dialog, modal, etc...
-      if ((currentRoute is PopupRoute ||
-              currentRoute is DialogRoute ||
-              currentRoute is RawDialogRoute ||
-              currentRoute is ModalBottomSheetRoute ||
-              currentRoute is CupertinoModalPopupRoute) &&
-          navigator.canPop()) {
-        // Pop the current dialog, modal, etc...
+      if (usePopUntil) {
+        return navigator.popUntil((route) => route is! PopupRoute);
+      }
+      if (modalRoute() is PopupRoute && navigator.canPop()) {
         navigator.pop();
-        // Recursively call this function to pop any other dialogs or modal sheets
-        dismissActivePopup();
+        dismissActivePopup(usePopUntil: usePopUntil);
       }
     } catch (_) {}
   }