+++ lib/src/value_notifier/notifier_classes/theme_mode_notifier.dart
@@ -11,6 +11,12 @@ extension ThemeModeNotifierEx on ValueNotifier<ThemeMode> {
   bool get isLight => value.isLight;
 
   bool get isSystem => value.isSystem;
+
+  void setDart() => value = ThemeMode.dark;
+
+  void setLight() => value = ThemeMode.light;
+
+  void setSystem() => value = ThemeMode.system;
 }
 
 class BrightnessNotifier extends ValueNotifier<Brightness> {
@@ -21,4 +27,8 @@ extension BrightnessNotifierEx on ValueNotifier<Brightness> {
   bool get isDark => value.isDark;
 
   bool get isLight => value.isLight;
+
+  void setDart() => value = Brightness.dark;
+
+  void setLight() => value = Brightness.light;
 }